chore: type 폴더 각 파일로 이동
